refactor: exitclose location
